refactor: DescartesCoordinateSystem is now in igraph.drawing.cairo
